+++ buildVars.py
@@ -54,7 +54,7 @@ https://github.com/linux-speakup/speakup"""),
 	# Add-on update channel (default is None, denoting stable releases,
 	# and for development releases, use "dev".)
 	# Do not change unless you know what you are doing!
-	addon_updateChannel="dev",
+	addon_updateChannel=None,
 	# Add-on license such as GPL 2
 	addon_license="GNU General Public License version 2 only",
 	# URL for the license document the ad-on is licensed under